+++ config/i18n_constants.php
@@ -1,0 +1,25 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * This file is part of me-tools.
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-tools
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
+ */
+
+define('I18N_ACTIONS', __d('me_tools', 'Actions'));
+define('I18N_DELETE', __d('me_tools', 'Delete'));
+define('I18N_EDIT', __d('me_tools', 'Edit'));
+define('I18N_NO', __d('me_tools', 'no'));
+define('I18N_OPERATION_OK', __d('me_tools', 'The operation has been performed correctly'));
+define('I18N_OPERATION_NOT_OK', __d('me_tools', 'The operation has not been performed correctly'));
+define('I18N_SUBMIT', __d('me_tools', 'Submit'));
+define('I18N_SURE_TO_DELETE', __d('me_tools', 'Are you sure you want to delete this?'));
+define('I18N_VIEW', __d('me_tools', 'View'));
+define('I18N_YES', __d('me_tools', 'yes'));